requirements.txt
pygetwindow
pyautogui
pynput
opencv-python
mss
paddleocr
paddlepaddle
langchain
chromadb
openai
pyyaml
python-dotenv
